[ALPS05269522] apusys power: fix ubsan build error
[Detail]
force casting to unsigned int to prevent
checker error in different compiler.

MTK-Commit-Id: 7f599b20f0d2ba33ab21b81b2cae86d2dc163324

CR-Id: ALPS05269522
Feature: [Module] APUSYS

+++ drivers/misc/mediatek/apusys/power/2.0/mt6885/hal_config_power.c
@@ -1142,7 +1142,8 @@ static int set_power_mtcmos(enum DVFS_USER user, void *param)
 				DRV_WriteReg32(APU_RPC_SW_FIFO_WE,
 					(domain_idx | BIT(4)));
 				LOG_DBG("%s APU_RPC_SW_FIFO_WE write 0x%x\n",
-					__func__, (domain_idx | BIT(4)));
+					__func__,
+					(unsigned int)(domain_idx | BIT(4)));
 
 				if (retry >= 3) {
 					LOG_ERR("%s fail (user:%d, mode:%d)\n",
